Ndrakulic/xla submodule (#16)
Added submodule tt-xla
Updated ./build_frontends.sh
- added building tt-mlir
- added building tt-xla
- Updated ./activate_frontend.sh
- Added activation of tt-xla
- deactivated previos enviroment if there is any
Additional fixes:
- At the end of ffe build unlinked ttmlir-toolchain so both frontends can be built at the same time.
- Check if ttforge-toolchain is already linked
- Added installing thomas package at the end of xla build

+++ build_frontends.sh
@@ -2,33 +2,71 @@
 # -o pipefail return error code from any command in a pipeline
 set -eo pipefail
 
-install_thomas() {
+install_tt_thomas() {
     pip install -e "$TT_FORGE_FE_HOME"
     pip install -r "$TT_FORGE_FE_HOME/requirements.txt"
 }
 
-build_forge_fe_env() {
+build_tt_forge_fe_env() {
     echo "Building forge frontend environment"
     source "$TT_FORGE_FE_HOME/env/activate"
     cmake -B "$TT_FORGE_FE_HOME/env/build" "$TT_FORGE_FE_HOME/env"
     cmake --build "$TT_FORGE_FE_HOME/env/build"
 }
 
-build_forge_fe() {
+build_tt_forge_fe() {
     echo "Building forge frontend"
     source "$TT_FORGE_FE_HOME/env/activate"
     cmake -G Ninja -B "$TT_FORGE_FE_HOME/build" "$TT_FORGE_FE_HOME"
     cmake --build "$TT_FORGE_FE_HOME/build"
 }
 
+build_tt_xla() {
+    echo "Building tt-xla"
+    if [ -z "$TTMLIR_TOOLCHAIN_DIR" ]; then
+        echo "TTMLIR_TOOLCHAIN_DIR is not set"
+        exit 1
+    fi
+
+    TT_XLA_HOME="$TT_THOMAS_HOME/third_party/tt-xla"
+    cd "$TT_XLA_HOME"
+    source venv/activate
+    cmake -G Ninja -B build
+    cmake --build build
+    cd "$TT_THOMAS_HOME"
+}
+
+build_tt_mlir() {
+    echo "Building tt-mlir"
+    if [ -z "$TTMLIR_TOOLCHAIN_DIR" ]; then
+        echo "TTMLIR_TOOLCHAIN_DIR is not set"
+        exit 1
+    fi
+
+    if [ -d "$TOOLCHAIN_DIR/tt-mlir" ]; then
+        rm -rf "$TOOLCHAIN_DIR/tt-mlir/"
+    fi
+
+    git clone https://github.com/tenstorrent/tt-mlir.git "$TOOLCHAIN_DIR/tt-mlir"
+    TT_MLIR_HOME="$TOOLCHAIN_DIR/tt-mlir"
+
+    mkdir -p "$TOOLCHAIN_DIR/tt-mlir/ttmlir-toolchain"
+    sudo ln -s "$TOOLCHAIN_DIR/tt-mlir/ttmlir-toolchain" /opt/
+
+    cmake -B "$TT_MLIR_HOME/env/build" "$TT_MLIR_HOME/env"
+    cmake --build "$TT_MLIR_HOME/env/build"
+}
+
 export TT_THOMAS_HOME="$(pwd)"
 
-build_tt_forge_fe=false
+tt_forge_fe=false
+tt_xla=false
 full_build=false
 
 while [[ "$#" -gt 0 ]]; do
     case $1 in
-        --ffe) build_tt_forge_fe=true ;;
+        --ffe) tt_forge_fe=true ;;
+        --xla) tt_xla=true ;;
         --full) full_build=true ;;
         *) echo "Unknown parameter passed: $1"; exit 1 ;;
     esac
@@ -37,7 +75,7 @@ done
 
 # Set the toolchain directory
 # This directory is used to store the toolchains for the different frontends
-if [ ! -v TOOLCHAIN_DIR ]; then
+if [ -z "$TOOLCHAIN_DIR" ]; then
     TOOLCHAIN_DIR="$TT_THOMAS_HOME/third_party/toolchains"
 fi
 
@@ -57,7 +95,7 @@ if ! command -v ninja &> /dev/null; then
     sudo apt install ninja-build
 fi
 
-if [ "$build_tt_forge_fe" = true ]; then
+if [ "$tt_forge_fe" = true ]; then
     export TT_FORGE_FE_HOME="$TT_THOMAS_HOME/third_party/tt-forge-fe"
 
     mkdir -p "$TOOLCHAIN_DIR/tt-forge-fe/ttforge-toolchain"
@@ -72,10 +110,35 @@ if [ "$build_tt_forge_fe" = true ]; then
     sudo ln -s "$TOOLCHAIN_DIR/tt-forge-fe/ttforge-toolchain" /opt/
 
     if [ "$full_build" = true ]; then        
-        build_forge_fe_env
+        build_tt_forge_fe_env
     fi
-    build_forge_fe
+    build_tt_forge_fe
 
-    install_thomas
+    install_tt_thomas
+    sudo unlink /opt/ttmlir-toolchain
+fi
+
+if [ "$tt_xla" = true ]; then
+    export TT_XLA_HOME="$TT_THOMAS_HOME/third_party/tt-xla"
+
+
+    # Fist we need to set TTMLIR_TOOLCHAIN_DIR
+    export TTMLIR_TOOLCHAIN_DIR="$OPT_MLIR_TOOLCHAIN_DIR"
+
+    if [ "$full_build" = true ]; then
+        build_tt_mlir
+        sudo unlink /opt/ttmlir-toolchain
+
+        if [ -d "$TOOLCHAIN_DIR/tt-xla" ]; then
+            rm -rf "$TOOLCHAIN_DIR/tt-xla/"
+        fi
+
+        mkdir -p "$TOOLCHAIN_DIR/tt-xla"
+        cp -r "$TOOLCHAIN_DIR/tt-mlir/ttmlir-toolchain" "$TOOLCHAIN_DIR/tt-xla/"
+    fi
+    sudo ln -s "$TOOLCHAIN_DIR/tt-xla/ttmlir-toolchain" /opt/
+
+    build_tt_xla
+    install_tt_thomas
     sudo unlink /opt/ttmlir-toolchain
 fi
